Refactor resizing logic to ConversationResizableContainer
